Implement 'emoji checker' example
Following an example from Microsoft [[1]].

This sets up the extension to run automatically against markdown files
and scans the file for uses of the word 'emoji'. If it finds one it
creates a `vscode.Diagnostic` and passes it back up to the extension
context.

This adds an underline to the text in the file which can be hovered over
to display some help text.

Hopefully this example sets up most of the boilerplate I need to display
error and warning messages for writing style.

[1]: https://github.com/microsoft/vscode-extension-samples/blob/main/code-actions-sample