Only run against Markdown files
The activation event for this extension is `"onLanguage:markdown"` which
means it's only started when a Markdown file is opened. Once it's
started though, it runs against ever file that's open in the session.

This isn't what we want - we only want to check Markdown files.

Early return if the document supplied isn't a Markdown file.